* Allow NONE as project classifier. * Add frontend support for so far missing project classifiers supported by backend. * Move available classifiers & collection logics into central mixin to reduce code duplication * Ensure classifiers are always sorted the same way in all views
