+++ src/i18n/locales/de.json
@@ -377,9 +377,11 @@
     "component_container": "Container",
     "component_cpe_desc": "Die CPE v2.2- oder v2.3-URI, wie von MITRE oder NIST bereitgestellt. Für alle Assets (Anwendungen, Betriebssysteme und Hardware) sollte ein CPE angegeben werden",
     "component_created": "Komponente erstellt",
+    "component_data": "Daten",
     "component_deleted": "Komponente gelöscht",
     "component_details": "Komponentendetails",
     "component_device": "Gerät",
+    "component_device_driver": "Gerätetreiber",
     "component_file": "Datei",
     "component_filename_desc": "Gibt den beobachteten Dateinamen der Komponente an",
     "component_firmware": "Firmware",
@@ -391,11 +393,14 @@
     "component_library": "Bibliothek",
     "component_license_expression_desc": "Gibt Lizenzinformationen für die Komponente in Form eines SPDX-Ausdrucks an",
     "component_license_url_desc": "Gibt die URL zur Lizenz der Komponente an",
+    "component_machine_learning_model": "Modell des maschinellen Lernens",
     "component_name": "Komponentenname",
     "component_name_desc": "Der vom Lieferanten angegebene Name der Komponente",
     "component_namespace_group_vendor": "Namespace/Gruppe/Anbieter",
+    "component_none_classifier": "Keiner",
     "component_operating_system": "Betriebssystem",
     "component_package_url_desc": "Für Bibliotheken und Frameworks ist eine gültige Paket-URL erforderlich. PURL-Syntax: pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "Plattform",
     "component_properties": "Komponenteneigenschaften",
     "component_search": "Komponentensuche",
     "component_spdx_license_desc": "Gibt die SPDX-Lizenz-ID der Komponente an",
