+++ src/i18n/locales/en.json
@@ -377,9 +377,11 @@
     "component_container": "Container",
     "component_cpe_desc": "The CPE v2.2 or v2.3 URI as provided by MITRE or NIST. All assets (applications, operating systems, and hardware) should have a CPE specified",
     "component_created": "Component created",
+    "component_data": "Data",
     "component_deleted": "Component deleted",
     "component_details": "Component Details",
     "component_device": "Device",
+    "component_device_driver": "Device Driver",
     "component_file": "File",
     "component_filename_desc": "Specifies the observed filename of the component",
     "component_firmware": "Firmware",
@@ -391,11 +393,14 @@
     "component_library": "Library",
     "component_license_expression_desc": "Specifies license information for the component in the form of an SPDX expression",
     "component_license_url_desc": "Specifies the URL to the license of the component",
+    "component_machine_learning_model": "Machine Learning Model",
     "component_name": "Component name",
     "component_name_desc": "The name of the component as provided by the supplier",
     "component_namespace_group_vendor": "Namespace / group / vendor",
+    "component_none_classifier": "None",
     "component_operating_system": "Operating system",
     "component_package_url_desc": "A Valid Package URL is required for libraries and frameworks. PURL syntax: pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "Platform",
     "component_properties": "Component Properties",
     "component_search": "Component Search",
     "component_spdx_license_desc": "Specifies the SPDX license ID of the component",
