+++ src/i18n/locales/es.json
@@ -377,9 +377,11 @@
     "component_container": "Contenedor",
     "component_cpe_desc": "El URI de CPE v2.2 o v2.3 proporcionado por MITRE o NIST. Todos los activos (aplicaciones, sistemas operativos y hardware) deben tener un CPE especificado.",
     "component_created": "Componente creado",
+    "component_data": "Datos",
     "component_deleted": "Componente eliminado",
     "component_details": "Detalles del componente",
     "component_device": "Dispositivo",
+    "component_device_driver": "Controlador de dispositivo",
     "component_file": "Archivo",
     "component_filename_desc": "Especifica el nombre de archivo observado del componente.",
     "component_firmware": "Firmware",
@@ -391,11 +393,14 @@
     "component_library": "Biblioteca",
     "component_license_expression_desc": "Especifica información de licencia para el componente en forma de expresión SPDX.",
     "component_license_url_desc": "Especifica la URL de la licencia del componente.",
+    "component_machine_learning_model": "Modelo de aprendizaje automático",
     "component_name": "Nombre del componente",
     "component_name_desc": "El nombre del componente proporcionado por el proveedor.",
     "component_namespace_group_vendor": "Espacio de nombres/grupo/proveedor",
+    "component_none_classifier": "Ninguno",
     "component_operating_system": "Sistema operativo",
     "component_package_url_desc": "Se requiere una URL de paquete válida para bibliotecas y marcos. Sintaxis PURL: paquete:tipo/espacio de nombres/nombre@versión?calificadores#subruta",
+    "component_platform": "Plataforma",
     "component_properties": "Propiedades de los componentes",
     "component_search": "Búsqueda de componentes",
     "component_spdx_license_desc": "Especifica el ID de licencia SPDX del componente.",
