+++ src/i18n/locales/fr.json
@@ -377,9 +377,11 @@
     "component_container": "Conteneur",
     "component_cpe_desc": "L'URI CPE v2.2 ou v2.3 tel que fourni par MITRE ou NIST. Tous les actifs (applications, systèmes d'exploitation et matériels) doivent avoir un CPE spécifié",
     "component_created": "Composant créé",
+    "component_data": "Données",
     "component_deleted": "Composant supprimé",
     "component_details": "Détails des composants",
     "component_device": "Équipement",
+    "component_device_driver": "Pilote de périphérique",
     "component_file": "Fichier",
     "component_filename_desc": "Spécifie le nom de fichier observé du composant",
     "component_firmware": "Microprogramme",
@@ -391,11 +393,14 @@
     "component_library": "Bibliothèque",
     "component_license_expression_desc": "Spécifie les informations de licence pour le composant sous la forme d'une expression SPDX",
     "component_license_url_desc": "Spécifie l'URL de la licence du composant",
+    "component_machine_learning_model": "Modèle d'apprentissage automatique",
     "component_name": "Nom du composant",
     "component_name_desc": "Le nom du composant tel que fourni par le fournisseur",
     "component_namespace_group_vendor": "Espace de noms/groupe/fournisseur",
+    "component_none_classifier": "Aucun",
     "component_operating_system": "Système d'exploitation",
     "component_package_url_desc": "Une PackageURL valide est requise pour les bibliothèques et les frameworks. Syntaxe PURL : pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "Plate-forme",
     "component_properties": "Propriétés des composants",
     "component_search": "Recherche de composants",
     "component_spdx_license_desc": "Spécifie l'ID de licence SPDX du composant",
