+++ src/i18n/locales/hi.json
@@ -377,9 +377,11 @@
     "component_container": "कंटेनर",
     "component_cpe_desc": "MITRE या NIST द्वारा प्रदान किया गया CPE v2.2 या v2.3 URI। सभी संपत्तियों (अनुप्रयोग, ऑपरेटिंग सिस्टम और हार्डवेयर) में CPE निर्दिष्ट होना चाहिए",
     "component_created": "घटक बनाया गया",
+    "component_data": "डेटा",
     "component_deleted": "घटक हटा दिया गया",
     "component_details": "घटक विवरण",
     "component_device": "डिवाइस",
+    "component_device_driver": "डिवाइस ड्राइवर",
     "component_file": "फ़ाइल",
     "component_filename_desc": "घटक का प्रेक्षित फ़ाइल नाम निर्दिष्ट करता है",
     "component_firmware": "फर्मवेयर",
@@ -391,11 +393,14 @@
     "component_library": "लाइब्रेरी",
     "component_license_expression_desc": "SPDX अभिव्यक्ति के रूप में घटक के लिए लाइसेंस जानकारी निर्दिष्ट करता है",
     "component_license_url_desc": "घटक के लाइसेंस का URL निर्दिष्ट करता है",
+    "component_machine_learning_model": "मशीन लर्निंग मॉडल",
     "component_name": "घटक का नाम",
     "component_name_desc": "आपूर्तिकर्ता द्वारा प्रदान किया गया घटक का नाम",
     "component_namespace_group_vendor": "नामस्थान / समूह / विक्रेता",
+    "component_none_classifier": "कोई नहीं",
     "component_operating_system": "ऑपरेटिंग सिस्टम",
     "component_package_url_desc": "लाइब्रेरीज़ और फ़्रेमवर्क के लिए एक मान्य पैकेज URL आवश्यक है। PURL सिंटैक्स: pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "प्लैटफ़ॉर्म",
     "component_properties": "घटक गुण",
     "component_search": "घटक खोज",
     "component_spdx_license_desc": "घटक की SPDX लाइसेंस आईडी निर्दिष्ट करता है",
