+++ src/i18n/locales/it.json
@@ -377,9 +377,11 @@
     "component_container": "Contenitore",
     "component_cpe_desc": "L'URI CPE v2.2 o v2.3 fornito da MITRE o NIST. Per tutte le risorse (applicazioni, sistemi operativi e hardware) deve essere specificato un CPE",
     "component_created": "Componente creato",
+    "component_data": "Dati",
     "component_deleted": "Componente eliminato",
     "component_details": "Dettagli dei componenti",
     "component_device": "Dispositivo",
+    "component_device_driver": "Driver del dispositivo",
     "component_file": "File",
     "component_filename_desc": "Specifica il nome file osservato del componente",
     "component_firmware": "Firmware",
@@ -391,11 +393,14 @@
     "component_library": "Libreria",
     "component_license_expression_desc": "Specifica le informazioni sulla licenza per il componente sotto forma di espressione SPDX",
     "component_license_url_desc": "Specifica l'URL della licenza del componente",
+    "component_machine_learning_model": "Modello di apprendimento automatico",
     "component_name": "Nome del componente",
     "component_name_desc": "Il nome del componente fornito dal fornitore",
     "component_namespace_group_vendor": "Spazio dei nomi/gruppo/fornitore",
+    "component_none_classifier": "Nessuno",
     "component_operating_system": "Sistema operativo",
     "component_package_url_desc": "Per librerie e framework è richiesto un URL del pacchetto valido. Sintassi PURL: pkg:tipo/spazio dei nomi/nome@versione?qualificatori#sottopercorso",
+    "component_platform": "Piattaforma",
     "component_properties": "Proprietà del componente",
     "component_search": "Ricerca componente",
     "component_spdx_license_desc": "Specifica l'ID della licenza SPDX del componente",
