+++ src/i18n/locales/ja.json
@@ -377,9 +377,11 @@
     "component_container": "コンテナ",
     "component_cpe_desc": "MITREまたはNISTによって提供されるCPE v2.2またはv2.3 URI。すべての資産（アプリケーション、オペレーティングシステム、ハードウェア）にはCPEを指定する必要があります。",
     "component_created": "コンポーネントが作成されました",
+    "component_data": "データ",
     "component_deleted": "コンポーネントが削除されました",
     "component_details": "コンポーネントの詳細",
     "component_device": "デバイス",
+    "component_device_driver": "デバイスドライバー",
     "component_file": "ファイル",
     "component_filename_desc": "コンポーネントの監視ファイル名を指定します",
     "component_firmware": "ファームウェア",
@@ -391,11 +393,14 @@
     "component_library": "ライブラリ",
     "component_license_expression_desc": "SPDX式の形式でコンポーネントのライセンス情報を指定します",
     "component_license_url_desc": "コンポーネントのライセンスのURLを指定します",
+    "component_machine_learning_model": "機械学習モデル",
     "component_name": "コンポーネント名",
     "component_name_desc": "サプライヤーから提供されたコンポーネントの名前",
     "component_namespace_group_vendor": "名前空間 / グループ / ベンダー",
+    "component_none_classifier": "なし",
     "component_operating_system": "オペレーティング システム",
     "component_package_url_desc": "ライブラリとフレームワークには有効なパッケージ URL が必要です。PURL 構文: pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "プラットフォーム",
     "component_properties": "コンポーネントのプロパティ",
     "component_search": "コンポーネント検索",
     "component_spdx_license_desc": "コンポーネントのSPDXライセンスIDを指定します",
