+++ src/i18n/locales/pl.json
@@ -377,9 +377,11 @@
     "component_container": "Kontener",
     "component_cpe_desc": "Identyfikator URI CPE v2.2 lub v2.3 dostarczony przez MITRE lub NIST. Wszystkie zasoby (aplikacje, systemy operacyjne i sprzęt) powinny mieć określony CPE",
     "component_created": "Komponent został utworzony",
+    "component_data": "Dane",
     "component_deleted": "Komponent usunięty",
     "component_details": "Szczegóły komponentu",
     "component_device": "Urządzenie",
+    "component_device_driver": "Sterownik urządzenia",
     "component_file": "Plik",
     "component_filename_desc": "Określa obserwowaną nazwę pliku komponentu",
     "component_firmware": "Oprogramowanie sprzętowe",
@@ -391,11 +393,14 @@
     "component_library": "Biblioteka",
     "component_license_expression_desc": "Określa informacje licencyjne dla komponentu w formie wyrażenia SPDX",
     "component_license_url_desc": "Określa adres URL licencji komponentu",
+    "component_machine_learning_model": "Model uczenia maszynowego",
     "component_name": "Nazwa komponentu",
     "component_name_desc": "Nazwa komponentu podana przez dostawcę",
     "component_namespace_group_vendor": "Przestrzeń nazw / grupa / dostawca",
+    "component_none_classifier": "Nic",
     "component_operating_system": "System operacyjny",
     "component_package_url_desc": "W przypadku bibliotek i struktur wymagany jest prawidłowy adres URL pakietu. Składnia PURL: pkg:typ/przestrzeń nazw/nazwa@wersja?kwalifikatory#podścieżka",
+    "component_platform": "Platforma",
     "component_properties": "Właściwości komponentu",
     "component_search": "Wyszukiwanie komponentów",
     "component_spdx_license_desc": "Określa identyfikator licencji SPDX komponentu",
