+++ src/i18n/locales/pt-BR.json
@@ -377,9 +377,11 @@
     "component_container": "Contêiner",
     "component_cpe_desc": "O URI do CPE v2.2 ou v2.3 conforme fornecido pelo MITRE ou NIST. Todos os ativos (aplicativos, sistemas operacionais e hardware) devem ter um CPE especificado",
     "component_created": "Componente criado",
+    "component_data": "Dados",
     "component_deleted": "Componente excluído",
     "component_details": "Detalhes do componente",
     "component_device": "Dispositivo",
+    "component_device_driver": "Driver de dispositivo",
     "component_file": "Arquivo",
     "component_filename_desc": "Especifica o nome do arquivo observado do componente",
     "component_firmware": "Firmware",
@@ -391,11 +393,14 @@
     "component_library": "Biblioteca",
     "component_license_expression_desc": "Especifica informações de licença do componente na forma de uma expressão SPDX",
     "component_license_url_desc": "Especifica o URL para a licença do componente",
+    "component_machine_learning_model": "Modelo de aprendizado de máquina",
     "component_name": "Nome do componente",
     "component_name_desc": "O nome do componente conforme fornecido pelo fornecedor",
     "component_namespace_group_vendor": "Namespace/grupo/fornecedor",
+    "component_none_classifier": "Nenhum",
     "component_operating_system": "Sistema operacional",
     "component_package_url_desc": "Um URL de pacote válido é necessário para bibliotecas e estruturas. Sintaxe PURL: pkg:tipo/namespace/nome@versão?qualificadores#subcaminho",
+    "component_platform": "Plataforma",
     "component_properties": "Propriedades do Componente",
     "component_search": "Pesquisa de Componentes",
     "component_spdx_license_desc": "Especifica o ID de licença SPDX do componente",
