+++ src/i18n/locales/ru.json
@@ -377,9 +377,11 @@
     "component_container": "Контейнер",
     "component_cpe_desc": "URI CPE v2.2 или v2.3, предоставленный MITRE или NIST. Для всех активов (приложений, операционных систем и оборудования) должно быть указано CPE.",
     "component_created": "Компонент создан",
+    "component_data": "Данные",
     "component_deleted": "Компонент удален.",
     "component_details": "Детали компонента",
     "component_device": "Устройство",
+    "component_device_driver": "Драйвер устройства",
     "component_file": "Файл",
     "component_filename_desc": "Указывает наблюдаемое имя файла компонента.",
     "component_firmware": "Прошивка",
@@ -391,11 +393,14 @@
     "component_library": "Библиотека",
     "component_license_expression_desc": "Указывает информацию о лицензии для компонента в форме выражения SPDX.",
     "component_license_url_desc": "Указывает URL-адрес лицензии компонента.",
+    "component_machine_learning_model": "Модель машинного обучения",
     "component_name": "Имя компонента",
     "component_name_desc": "Название компонента, предоставленное поставщиком",
     "component_namespace_group_vendor": "Пространство имен/группа/поставщик",
+    "component_none_classifier": "Никто",
     "component_operating_system": "Операционная система",
     "component_package_url_desc": "Для библиотек и платформ требуется действительный URL-адрес пакета. Синтаксис PURL: pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "Платформа",
     "component_properties": "Свойства компонента",
     "component_search": "Поиск компонентов",
     "component_spdx_license_desc": "Указывает идентификатор лицензии SPDX компонента.",
