+++ src/i18n/locales/uk-UA.json
@@ -377,9 +377,11 @@
     "component_container": "Контейнер",
     "component_cpe_desc": "URI CPE версії 2.2 або версії 2.3, наданий МІТЕР або NIST. \nУсі активи (додатки, операційні системи та апаратне забезпечення) повинні мати CPE",
     "component_created": "Компонент створено",
+    "component_data": "Дані",
     "component_deleted": "Компонент видалено",
     "component_details": "Деталі компонента",
     "component_device": "пристрій",
+    "component_device_driver": "Драйвер пристрою",
     "component_file": "Файл",
     "component_filename_desc": "Визначає спостережуване ім'я файлу компонента",
     "component_firmware": "Прошивка",
@@ -391,11 +393,14 @@
     "component_library": "Бібліотека",
     "component_license_expression_desc": "Вказує інформацію про ліцензію для компонента у формі виразу SPDX",
     "component_license_url_desc": "Вказує URL-адресу ліцензії компонента",
+    "component_machine_learning_model": "Модель машинного навчання",
     "component_name": "Назва компонента",
     "component_name_desc": "Назва компонента, надана постачальником",
     "component_namespace_group_vendor": "Простір імен/група/постачальник",
+    "component_none_classifier": "Жодного",
     "component_operating_system": "Операційна система",
     "component_package_url_desc": "Для бібліотек і фреймворків потрібна дійсна URL-адреса пакета. \nСинтаксис PURL: pkg:type/namespace/name@version?qualifiers",
+    "component_platform": "Платформа",
     "component_properties": "Властивості компонента",
     "component_search": "Пошук компонентів",
     "component_spdx_license_desc": "Визначає ідентифікатор ліцензії SPDX компонента",
