+++ src/i18n/locales/zh.json
@@ -377,9 +377,11 @@
     "component_container": "容器",
     "component_cpe_desc": "MITRE 或 NIST 提供的 CPE v2.2 或 v2.3 URI。所有资产（应用程序、操作系统和硬件）都应指定 CPE",
     "component_created": "组件已创建",
+    "component_data": "数据",
     "component_deleted": "组件已删除",
     "component_details": "组件详细信息",
     "component_device": "设备",
+    "component_device_driver": "设备驱动程序",
     "component_file": "文件",
     "component_filename_desc": "指定观察到的组件的文件名",
     "component_firmware": "固件",
@@ -391,11 +393,14 @@
     "component_library": "库",
     "component_license_expression_desc": "以 SPDX 表达式的形式指定组件的许可证信息",
     "component_license_url_desc": "指定组件许可证的 URL",
+    "component_machine_learning_model": "机器学习模型",
     "component_name": "组件名称",
     "component_name_desc": "供应商提供的组件名称",
     "component_namespace_group_vendor": "命名空间/组/供应商",
+    "component_none_classifier": "没有任何",
     "component_operating_system": "操作系统",
     "component_package_url_desc": "库和框架需要有效的包 URL。PURL 语法：pkg:type/namespace/name@version?qualifiers#subpath",
+    "component_platform": "平台",
     "component_properties": "组件属性",
     "component_search": "组件搜索",
     "component_spdx_license_desc": "指定组件的 SPDX 许可证 ID",
