+++ src/views/portfolio/projects/ComponentDetailsModal.vue
@@ -125,7 +125,7 @@
             id="component-classifier-input"
             required="true"
             v-model="component.classifier"
-            :options="availableClassifiers"
+            :options="sortAvailableClassifiers"
             :label="$t('message.classifier')"
             :tooltip="$t('message.component_classifier_desc')"
             :disabled="this.isNotPermitted(PERMISSIONS.PORTFOLIO_MANAGEMENT)"
@@ -391,10 +391,11 @@ import ComponentPropertiesModal from './ComponentPropertiesModal.vue';
 import permissionsMixin from '../../../mixins/permissionsMixin';
 import xssFilters from 'xss-filters';
 import common from '@/shared/common';
+import availableClassifiersMixin from '@/mixins/availableClassifiersMixin';
 
 export default {
   name: 'ComponentDetailsModal',
-  mixins: [permissionsMixin],
+  mixins: [permissionsMixin, availableClassifiersMixin],
   components: {
     BInputGroupFormInput,
     BInputGroupFormSelect,
@@ -405,28 +406,6 @@ export default {
   },
   data() {
     return {
-      availableClassifiers: [
-        {
-          value: 'APPLICATION',
-          text: this.$i18n.t('message.component_application'),
-        },
-        {
-          value: 'FRAMEWORK',
-          text: this.$i18n.t('message.component_framework'),
-        },
-        { value: 'LIBRARY', text: this.$i18n.t('message.component_library') },
-        {
-          value: 'CONTAINER',
-          text: this.$i18n.t('message.component_container'),
-        },
-        {
-          value: 'OPERATING_SYSTEM',
-          text: this.$i18n.t('message.component_operating_system'),
-        },
-        { value: 'DEVICE', text: this.$i18n.t('message.component_device') },
-        { value: 'FIRMWARE', text: this.$i18n.t('message.component_firmware') },
-        { value: 'FILE', text: this.$i18n.t('message.component_file') },
-      ],
       selectableLicenses: [],
       selectedLicense: '',
       supplierUrlsTableColumns: [
